+++ packages/x6/src/common/dom-event/constant.ts
@@ -1,0 +1,8 @@
+export const CLICK = 'click'
+export const DOUBLE_CLICK = 'doubleClick'
+export const MOUSE_DOWN = 'mouseDown'
+export const MOUSE_MOVE = 'mouseMove'
+export const MOUSE_UP = 'mouseUp'
+export const ESCAPE = 'escape'
+export const FIRE_MOUSE_EVENT = 'fireMouseEvent'
+export const START_EDITING = 'startEditing'
