+++ packages/x6/src/common/dom-event/index.ts
@@ -1,0 +1,4 @@
+import * as DomEvent from './index.rollup'
+
+export { DomEvent }
+export * from './mouse'
