+++ packages/x6/src/common/index.ts
@@ -1,8 +1,7 @@
+export * from './detector'
 export * from './events'
 export * from './disposable'
 export * from './primer'
 export * from './disablable'
-export * from './detector'
-export * from './dom-event'
-export * from './mouse-event'
 export * from './dictionary'
+export * from './dom-event'
